Added Camera Interactivity
Slightly tweaked the vertex shader to include position. Used a series of event listeners to check for mouse movement. Once detected, a function is run using a 3x3 matrix to update the zoom and pan. This makes use of a homogeneous coordinate system to speed up computing time by handling scale and location at the same time.  Then, each frame, this matrix is retrieved and parsed to WebGL to update the graphics display.