Switched Physics Calculation System to Leapfrog Integration
Used leapfrog integration to update the velocity of bodies in smaller steps, preserving accuracy and preventing energy from increasing in the system. Added trial cut off by checking distance between the first quater of the trials of each body to see if they overlap with the current trial. This gives a more accurate and visually appealing trial and orbit system.