Added menu which includes a togglable 'glow' effect and a add new planet function.
The camera now moves while paused by updating the view matrix by requesting frames separately from the main loop if paused. Moved formatting into a separate CSS file to make room for the menu. The menu animates in when clicked and has hover-over instructions over the add new planet button. Added a new planet button that creates a new planet where the user clicks, and it gets bigger the longer the user holds the mouse down. The new planet can also be given velocity by moving the mouse. Also made a glow effect and button to toggle it, as it's rather laggy.