Added Support for Moons and Satellites
When the user hovers over the timespeed option, it warns them of the risks of speeding up the sim, namely, how it creates errors leading to instability. Created a new function that goes through every celestial body every 5 seconds to check which body it is orbiting by comparing the ratio of each sphere of influence. This changes the new attribute to the body, which is called parent. If a body has a parent, its trail positions are displayed relative to their parent. This is so if a planet is orbiting the sun, its moon will show its orbit around the planet rather than the strange zig-zag pattern. To prevent bugs when a planet is removed, the parent indicies of the over planets is shifted. To further strengthen the detection of what body the planet is orbiting, it uses a conditional operator to increase the gravitational influence of the body it was last known to orbit to prevent switching parents when there are small changes. This most importantly prevents moons from desyncing when they come close to a more massive body, such as the sun.