Added Camera Controls
The sim now features three buttons at the top of the screen. The left and right buttons switch through the list of bodies and lock and move the camera with the planets until the user drags the mouse. This triggers the camera to be released, and the user can then pan the camera however they want. The central button displays a list of all planets with their respective colours. When one of these options is selected, the camera then follows that planet.