Add Interactive Graphing
The user can now switch between different planets and then click the display graph, which shows the graphs of the velocity of that planet over time. It clears the results when the selected planet switches and automatically restarts timing to prevent the accumulation of unnecessary data, which slows down the simulation. The new graph system shows how velocity of planets ossilates, with them moving faster when closer to the sun - demonstrating kepler's second law of planetary motion.